+++ axelor-base/src/main/java/com/axelor/apps/base/ical/ICalendarService.java
@@ -27,7 +27,6 @@ import com.axelor.apps.base.exceptions.IExceptionMessage;
 import com.axelor.apps.base.service.app.AppBaseService;
 import com.axelor.apps.message.db.EmailAddress;
 import com.axelor.apps.message.db.repo.EmailAddressRepository;
-import com.axelor.apps.message.service.MailAccountService;
 import com.axelor.apps.tool.QueryBuilder;
 import com.axelor.auth.db.User;
 import com.axelor.common.StringUtils;
@@ -123,8 +122,6 @@ public class ICalendarService {
 
   @Inject protected ICalendarEventRepository iEventRepo;
 
-  @Inject private MailAccountService mailAccountService;
-
   public static class GenericPathResolver extends PathResolver {
 
     private String principalPath;
@@ -190,7 +187,7 @@ public class ICalendarService {
     ICalendarStore store = new ICalendarStore(url, RESOLVER);
 
     try {
-      store.connect(cal.getLogin(), getCalendarDecryptPassword(cal.getPassword()));
+      store.connect(cal.getLogin(), cal.getPassword());
     } finally {
       store.disconnect();
     }
@@ -627,7 +624,7 @@ public class ICalendarService {
     URL url = new URL(protocol.getScheme(), calendar.getUrl(), calendar.getPort(), "");
     ICalendarStore store = new ICalendarStore(url, RESOLVER);
     try {
-      String password = getCalendarDecryptPassword(calendar.getPassword());
+      String password = calendar.getPassword();
 
       if (calendar.getLogin() != null
           && calendar.getPassword() != null
@@ -927,8 +924,7 @@ public class ICalendarService {
       URL url = new URL(protocol.getScheme(), calendar.getUrl(), calendar.getPort(), "");
       ICalendarStore store = new ICalendarStore(url, RESOLVER);
       try {
-        if (store.connect(
-            calendar.getLogin(), getCalendarDecryptPassword(calendar.getPassword()))) {
+        if (store.connect(calendar.getLogin(), calendar.getPassword())) {
           List<CalDavCalendarCollection> colList = store.getCollections();
           if (!colList.isEmpty()) {
             CalDavCalendarCollection collection = colList.get(0);
@@ -977,15 +973,5 @@ public class ICalendarService {
         .all()
         .filter("COALESCE(self.archived, false) = false AND self.calendar = ?1", calendar)
         .fetch();
-  }
-
-  public String getCalendarEncryptPassword(String password) {
-
-    return mailAccountService.getEncryptPassword(password);
-  }
-
-  public String getCalendarDecryptPassword(String password) {
-
-    return mailAccountService.getDecryptPassword(password);
   }
 }