+++ axelor-message/src/main/java/com/axelor/apps/message/service/MailAccountService.java
@@ -40,10 +40,6 @@ public interface MailAccountService {
 
   public String getSignature(EmailAccount mailAccount);
 
-  public String getEncryptPassword(String password);
-
-  public String getDecryptPassword(String password);
-
   public int fetchEmails(EmailAccount mailAccount, boolean unseenOnly)
       throws MessagingException, IOException;